# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Category {
  id: number;
  nameAr: string;
  slug: string;
  imageUrl: string;
}

export interface Variant {
  id: number;
  nameAr: string;
  hex: string;
  sku: string;
  stock: number;
}

export interface Product {
  id: number;
  nameAr: string;
  slug: string;
  descriptionAr: string;
  price: number;
  /** @nullable */
  compareAtPrice?: number | null;
  category: string;
  imageUrl: string;
  rating: number;
  reviewCount: number;
  /** @nullable */
  badge: string | null;
  variants: Variant[];
}

export interface Testimonial {
  id: number;
  name: string;
  quote: string;
  rating: number;
}

export interface StorefrontSummary {
  categories: Category[];
  featuredProducts: Product[];
  testimonials: Testimonial[];
}

export interface ShippingRateInput {
  country: string;
  /** @minimum 1 */
  weightGrams: number;
}

export interface ShippingRate {
  id: string;
  carrier: string;
  service: string;
  price: number;
  estimatedDays: string;
}

export interface OrderItemInput {
  productId: number;
  /** @nullable */
  variantId?: number | null;
  /** @minimum 1 */
  quantity: number;
}

export interface OrderInput {
  items: OrderItemInput[];
  email: string;
  shippingAddress: string;
}

export interface Order {
  id: number;
  status: string;
  totalAmount: number;
  createdAt: string;
}

export type ListProductsParams = {
category?: string;
search?: string;
featured?: boolean;
};