# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItemInput } from './orderItemInput';

export interface OrderInput {
  items: OrderItemInput[];
  email: string;
  shippingAddress: string;
}