# lib/api-zod/src/generated/types/orderItemInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OrderItemInput {
  productId: number;
  /** @nullable */
  variantId?: number | null;
  /** @minimum 1 */
  quantity: number;
}