# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Variant } from './variant';

export interface Product {
  id: number;
  nameAr: string;
  slug: string;
  descriptionAr: string;
  price: number;
  /** @nullable */
  compareAtPrice?: number | null;
  category: string;
  imageUrl: string;
  rating: number;
  reviewCount: number;
  /** @nullable */
  badge: string | null;
  variants: Variant[];
}