# lib/api-zod/src/generated/types/shippingRate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ShippingRate {
  id: string;
  carrier: string;
  service: string;
  price: number;
  estimatedDays: string;
}