# lib/api-zod/src/generated/types/shippingRateInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ShippingRateInput {
  country: string;
  /** @minimum 1 */
  weightGrams: number;
}