# lib/api-zod/src/generated/types/storefrontSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Category } from './category';
import type { Product } from './product';
import type { Testimonial } from './testimonial';

export interface StorefrontSummary {
  categories: Category[];
  featuredProducts: Product[];
  testimonials: Testimonial[];
}