# lib/api-zod/src/generated/types/testimonial.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Testimonial {
  id: number;
  name: string;
  quote: string;
  rating: number;
}